Fix heuristic error in bestfs

+++ src/bestfs.h
@@ -3,7 +3,7 @@
 
 #include "util.h"
 
-int heuristic(int node, int dest, const vector<pair<float,float>>& coordinates);
+double heuristic(int node, int dest, const vector<pair<float,float>>& coordinates);
 pair<vector<int>,int> bestFS(const vector<vector<int>>& adjList, const vector<pair<float,float>>& coordinates, int src, int dest);
 
 #endif
